backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
inference-sdk>=0.16.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.4
pillow==10.1.0